Separated telemetry parts in traces and metrics
